powernet concurrency

+++ src/main/java/com/aivech/tau/power/IRotaryProvider.java
@@ -1,6 +1,8 @@
 package com.aivech.tau.power;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 public interface IRotaryProvider extends IRotaryBlock {
-    long getSpeed();
-    long getTorque();
+    AtomicInteger getSpeed();
+    AtomicInteger getTorque();
 }
